pyproject.toml
[project]
name = "wyoming-universal-stt"
version = "0.1.0"
description = "Wyoming Server for Faster Whisper"
readme = "README.md"
requires-python = ">=3.8.1,<3.12"
license = {text = "MIT"}
authors = [
    {name = "Rushi Chaudhari"}
]
keywords = ["wyoming", "whisper", "stt"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Text Processing :: Linguistic",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "numpy>=1.24.4",
    "pyaudio>=0.2.14",
    "wyoming>=1.5.3",
]

[project.urls]
Homepage = "http://github.com/0xrushi/wyoming-universal-stt"

[project.scripts]
wyoming-universal-stt = "wyoming_universal_stt.__main__:run"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
check_untyped_defs = true
disallow_untyped_defs = true

[project.optional-dependencies]
faster-whisper = [
    "faster-whisper>=1.1.0"
]

mlx = [
    "mlx-whisper>=0.4.0"
]

openai-whisper = [
    "openai>=1.0.0"
]

all = [
    "faster-whisper>=1.1.0",
    "mlx-whisper>=0.1.0",
    "openai>=1.0.0"
]

dev = [
    "black==22.12.0",
    "flake8==6.0.0",
    "isort==5.11.3",
    "mypy==0.991",
    "pylint==2.15.9",
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3"
]
